create utils.py
