+++ flask_app.py
@@ -3,7 +3,6 @@ import subprocess
 from shlex import quote
 import json
 # сериализовать выход парсера
-import re
 
 from flask import (
     Flask,
@@ -20,18 +19,11 @@ from flask import (
 )
 from werkzeug.utils import secure_filename
 from dicttoxml import dicttoxml
+from utils import parse_hfst, TRANSDUCER_MAPPING, is_allowed_extension
 
 app = Flask(__name__)
 
 APP_ROOT = os.path.dirname(os.path.abspath(__file__))
-
-ALLOWED_EXTENSIONS = {'txt'}
-
-
-def is_allowed_extension(filename: str):
-    return any(
-        map(lambda x: filename.endswith("." + x), ALLOWED_EXTENSIONS) 
-    )
 
 
 @app.route("/")
@@ -44,68 +36,13 @@ def asr():
     return render_template("asr.html", title="FieldNLP ASR")
 
 
-def parse_hfst(hfst_out: str) -> list:
-    """turn hfst output into serializeable format"""
-    words = re.findall(r"(?<=\^).+?(?=\$)", hfst_out)
-    out_dicts = []
-    for word in words:
-        parts = word.split("/")
-        word_dict = {"token": parts[0], "analyses": []}
-        # if word is unrecognized, use placeholder
-        if parts[1].startswith("*"):
-            word_dict["analyses"].append(
-                {
-                    "lemma": "UNK",
-                    "gloss": "UNK"
-                }
-            )
-            out_dicts.append(word_dict)
-            continue
-        for part in parts[1:]:
-            if part.startswith("<"):
-                word_dict["analyses"].append(
-                    {
-                        "lemma": "" if not (x := re.search("(?<=\>)\w+(?=\<)", part)) else x.group(),
-                        "gloss": ".".join(
-                            re.findall(r"[^<>]+", part)
-                        )
-                    }
-                )
-            else:
-                breakpnt = part.find("<")
-                word_dict["analyses"].append(
-                    {
-                        "lemma": part[:breakpnt],
-                        # change gloss format
-                        "gloss": ".".join(
-                            re.findall(r"[^<>]+", part[breakpnt:])
-                        )
-                    }
-                )
-        out_dicts.append(word_dict)
-    return out_dicts
-
 @app.route("/parsers", methods=["GET", "POST"])
 def parsers():
-    TRANSDUCER_MAPPING = {
-        "abz": {
-            "html_name": "Абазинский",
-            "filename": "abaza.ana.hfstol",
-        },
-        "bgv-cyr": {
-            "html_name": "Багвалинский (аварский алфавит)",
-            "filename": "merged.hfstol",
-        },
-        "bgv-lat": {
-            "html_name": "Багвалинский (транскрипция)",
-            "filename": "merged.tr.hfstol"
-        }
-    }
     params = {
         "title": "FieldNLP Parsers",
         "request_text": "",
         "response_text": [],
-        "mapping":TRANSDUCER_MAPPING,
+        "mapping": TRANSDUCER_MAPPING,
         "curlang": "none"
     }
     if request.method == "GET":
